Reorganize ImGui menus
