+++ libs/JSystem/src/JFramework/JFWDisplay.cpp
@@ -379,6 +379,7 @@ static void waitPrecise(Limiter& limiter, Uint64 targetNs) {
 static void waitForTick(u32 p1, u16 p2) {
 #if TARGET_PC
     if (dusk::getSettings().game.enableFrameInterpolation && !dusk::getTransientSettings().skipFrameRateLimit) {
+        dusk::frameUsagePct = 0.f;
         return;
     }
     if (dusk::getTransientSettings().skipFrameRateLimit) {
