+++ src/dusk/imgui/ImGuiConsole.cpp
@@ -319,9 +319,11 @@ namespace dusk {
             }
         }
 
+        // The menu bar renders with ImGuiCol_WindowBg behind it. We just want ImGuiCol_MenuBarBg,
+        // so make the window bg fully transparent temporarily
+        ImGui::PushStyleColor(ImGuiCol_WindowBg, ImVec4(0.0f, 0.0f, 0.0f, 0.0f));
         if (showMenu && ImGui::BeginMainMenuBar()) {
             m_menuGame.draw();
-            m_menuEnhancements.draw();
             m_menuTools.draw();
 
             const auto fpsLabel =
@@ -336,6 +338,7 @@ namespace dusk {
 
             ImGui::EndMainMenuBar();
         }
+        ImGui::PopStyleColor();
 
         if (!getSettings().backend.wasPresetChosen) {
             m_firstRunPreset.draw();
