--- src/dusk/imgui/ImGuiMenuEnhancements.hpp
@@ -1,18 +1,0 @@
-#ifndef DUSK_IMGUI_MENUENHANCEMENTS_HPP
-#define DUSK_IMGUI_MENUENHANCEMENTS_HPP
-
-#include <aurora/aurora.h>
-#include <pad.h>
-#include <string>
-
-#include "imgui.h"
-
-namespace dusk {
-    class ImGuiMenuEnhancements {
-    public:
-        ImGuiMenuEnhancements();
-        void draw();
-    };
-}
-
-#endif  // DUSK_IMGUI_MENUENHANCEMENTS_HPP
